requirements.txt
python-docx==1.1.0
setuptools==70.0.0
wheel==0.43.0
pillow==10.3.0